requirements.txt

httpx>=0.27.0
pydantic>=2.7.0
pyyaml>=6.0.1
beautifulsoup4>=4.12.3
rich>=13.7.1
fastapi>=0.111.0
uvicorn>=0.30.0
orjson>=3.10.0
structlog>=24.1.0
psutil>=6.0.0

sentry-sdk>=2.5.0

dnspython>=2.6.1

tldextract>=5.1.2
